requirements-test.txt
nose>=1.3.0
django-nose>=1.2
flake8>=2.1.0
tox>=1.8.0
tox-pyenv
wheel
check-manifest
bleach<2.0
readme_renderer<17
django-coverage-plugin==1.5.0
pylint
feedparser
